requirements.txt
berserk==0.14.0
certifi==2025.11.12
charset-normalizer==3.4.4
chess==1.11.2
Deprecated==1.3.1
idna==3.11
ndjson==0.3.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.45
stockfish==4.0.3
typing_extensions==4.15.0
urllib3==2.6.2
wrapt==2.0.1
